move sensitive data to json config, add possibility to load different configs
