+++ youtrack.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 # -*- encoding: utf-8 -*-
-from config import YoutrackConfig
+from configuration import YoutrackConfig
 import requests
 from lxml import etree
